+++ netlify.toml
@@ -1,5 +1,3 @@
 [build]
   command = "npm run setup"
   publish = "build"
-[[plugins]]
-  package = "netlify-plugin-cypress"
