+++ package.json
@@ -2,7 +2,7 @@
   "name": "bookshelf",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://bookshelf.lol/",
+  "homepage": "https://exercises-10-render-as-you-fetch.bookshelf.lol/",
   "engines": {
     "node": "^10.13 || 12 || 14",
     "npm": ">=6"
@@ -77,7 +77,7 @@
     "format": "prettier --write \"**/*.+(js|json|css|md|mdx|html)\"",
     "setup": "node setup",
     "validate-exercises": "node ./scripts/validate-exercises",
-    "validate": "npm run validate-exercises && npm-run-all --parallel lint test:coverage build"
+    "validate": "npm run validate-exercises && npm-run-all --parallel lint test:coverage && node ./scripts/build"
   },
   "eslintConfig": {
     "extends": "react-app"
