--- src/setupTests.js
@@ -1,52 +1,0 @@
-import '@testing-library/jest-dom/extend-expect'
-import {configure, act} from '@testing-library/react'
-import {queryCache} from 'react-query'
-import * as auth from 'auth-provider'
-import {server} from 'test/server'
-import * as usersDB from 'test/data/users'
-import * as listItemsDB from 'test/data/list-items'
-import * as booksDB from 'test/data/books'
-
-// we don't need the profiler in tests
-jest.mock('components/profiler')
-
-// set the location to the /list route as we auto-redirect users to that route
-window.history.pushState({}, 'Home page', '/list')
-
-// speeds up *ByRole queries a bit
-// https://github.com/testing-library/dom-testing-library/issues/552
-configure({defaultHidden: true})
-
-// make debug output for TestingLibrary Errors larger
-process.env.DEBUG_PRINT_LIMIT = 15000
-
-// enable API mocking in test runs using the same request handlers
-// as for the client-side mocking.
-beforeAll(() => server.listen({onUnhandledRequest: 'error'}))
-afterAll(() => server.close())
-afterEach(() => server.resetHandlers())
-
-// general cleanup
-afterEach(async () => {
-  queryCache.clear()
-  await Promise.all([
-    auth.logout(),
-    usersDB.reset(),
-    booksDB.reset(),
-    listItemsDB.reset(),
-  ])
-})
-
-// real times is a good default to start, individual tests can
-// enable fake timers if they need, and if they have, then we should
-// run all the pending timers (in `act` because this can trigger state updates)
-// then we'll switch back to realTimers.
-// it's important this comes last here because jest runs afterEach callbacks
-// in reverse order and we want this to be run first so we get back to real timers
-// before any other cleanup
-afterEach(() => {
-  if (jest.isMockFunction(setTimeout)) {
-    act(() => jest.runOnlyPendingTimers())
-    jest.useRealTimers()
-  }
-})
